+++ src/ServiceFactory.php
@@ -4,17 +4,35 @@ namespace Wikibase\Api;
 
 use DataValues\Deserializers\DataValueDeserializer;
 use Mediawiki\Api\MediawikiApi;
+use Wikibase\Api\Service\RevisionRepo;
+use Wikibase\Api\Service\RevisionSaver;
 use Wikibase\DataModel\DeserializerFactory;
 use Wikibase\DataModel\Entity\BasicEntityIdParser;
 
-class RepositoryFactory {
+class ServiceFactory {
 
+	/**
+	 * @var MediawikiApi
+	 */
 	private $api;
 
+	/**
+	 * @param MediawikiApi $api
+	 */
 	public function __construct( MediawikiApi $api ) {
 		$this->api = $api;
 	}
 
+	/**
+	 * @return RevisionSaver
+	 */
+	public function newRevisionSaver() {
+		return new RevisionSaver( $this->api );
+	}
+
+	/**
+	 * @return RevisionRepo
+	 */
 	public function newRevisionRepo() {
 		return new RevisionRepo(
 			$this->api,