+++ RELEASENOTES.md
@@ -4,6 +4,8 @@ These are the release notes for the [wikibase-api](README.md).
 
 * Adjust for changed in mediawiki-api 0.2
 * Internal classes renamed
+* Introduced a Service namespace and ServiceFactory
+* Introduce Property and Item Content objects
 
 ## Version 0.1 (2014-02-23)
 
@@ -12,5 +14,5 @@ Initial release with the following features:
 * EntityRevision
 * RevisionRepo
 * RevisionSaver
-* RepositoryFactory
+* ServiceFactory
 * Can Get, Edit, Save, Create, Manipulate Entities